Add perm for limited access to registration entries

+++ src/workers/http/routing/registration/entries/$registrationEntryId/delete.js
@@ -1,16 +1,29 @@
 export default {
 	schema: {
 		requirePerms: [
-			'registration.entries.delete',
 			'codeholders.read',
 		],
 	},
 
 	run: async function run (req, res) {
-		const deleted = await AKSO.db('registration_entries')
+		// Make sure the registration entry exists
+		const registrationEntry = await AKSO.db('registration_entries')
+			.where('id', req.params.registrationEntryId)
+			.first('intermediary');
+		if (!registrationEntry) {
+			return res.sendStatus(404);
+		}
+
+		// Check perms
+		if (!req.hasPermission('registration.entries.delete') &&
+			!req.hasPermission(`registration.entries.intermediary.${registrationEntry.intermediary}`)) {
+			return res.sendStatus(403);
+		}
+
+		await AKSO.db('registration_entries')
 			.where('id', req.params.registrationEntryId)
 			.delete();
 
-		res.sendStatus(deleted ? 204 : 404);
+		res.sendStatus(204);
 	}
 };
